backend/requirements.txt
Django==4.2.9
# Use a currently available DRF release (3.16.1 is available on PyPI)
djangorestframework==3.16.1
python-dotenv==1.0.0
django-cors-headers==4.0.0
